requirements.txt
numpy
matplotlib
torch
pytorch-lightning
torchvision
torchmetrics
lightning
opencv-python
scikit-image
multiprocess
timm
einops
transformers
wandb
pyyaml
argparse
SimpleITK
pydicom
pylibjpeg
pylibjpeg-libjpeg
scipy
scikit-learn
itk
itk-rtk
pandas
h5py